requirement.txt
fastapi
uvicorn[standard]
email-validator
passlib
bcrypt
python-jose
cryptography
python-multipart
asyncio
websocket-client
pandas
APScheduler
motor
redis
pyyaml 
ua-parser 
user-agents
pyfcm
python-dotenv
websockets
PyPubSub
fastapi-mail
fastapi-events
pyotp